+++ index.js
@@ -4,25 +4,18 @@ module.exports = {
 	extends: ['stylelint-config-standard-scss', 'stylelint-config-suitcss'],
 	plugins: ['stylelint-high-performance-animation'],
 	rules: {
-		// redeclared from standard-scss because suitcss sets to `always`
-		'block-closing-brace-newline-after': [
-			'always',
-			{
-				ignoreAtRules: ['if', 'else'],
-			},
-		],
 		// these are being set in stylelint-standard, but we don't want them
 		'alpha-value-notation': null, // not ready for this syntax yet
 		'custom-media-pattern': null,
 		'custom-property-pattern': null,
 		'declaration-empty-line-before': null, // false errors after SCSS comments
 		'keyframes-name-pattern': null,
+		'selector-class-pattern': null,
+		'selector-id-pattern': null,
 		'scss/at-function-pattern': null,
 		'scss/at-mixin-pattern': null,
 		'scss/dollar-variable-pattern': null,
 		'scss/percent-placeholder-pattern': null,
-		'selector-class-pattern': null,
-		'selector-id-pattern': null,
 		// our rules from here on
 		'at-rule-disallowed-list': [
 			['extend', 'import'],
@@ -36,13 +29,6 @@ module.exports = {
 			'always',
 			{
 				except: ['first-nested'],
-			},
-		],
-		'max-line-length': [
-			80,
-			{
-				ignore: 'non-comments',
-				ignorePattern: '/https?://[0-9,a-z]*.*/',
 			},
 		],
 		'no-descending-specificity': null,